graph, chain/ethereum: Move subgraph component traits
