+++ chain/ethereum/src/types/mod.rs
@@ -1,0 +1,5 @@
+mod indexer;
+mod instance;
+
+pub use self::indexer::SubgraphIndexer;
+pub use self::instance::SubgraphInstance;
