+++ graph/src/components/subgraph/mod.rs
@@ -1,15 +1,11 @@
-mod host;
-mod indexer;
-mod instance;
 mod loader;
+mod metrics;
 mod provider;
 mod registrar;
 
 pub use crate::prelude::Entity;
 
-pub use self::host::{HostMetrics, RuntimeHost, RuntimeHostBuilder};
-pub use self::indexer::SubgraphIndexer;
-pub use self::instance::{BlockState, DataSourceTemplateInfo, SubgraphInstance};
 pub use self::loader::DataSourceLoader;
+pub use self::metrics::HostMetrics;
 pub use self::provider::SubgraphAssignmentProvider;
 pub use self::registrar::{SubgraphRegistrar, SubgraphVersionSwitchingMode};
